requirements.txt
numpy==1.24.0
scipy
pandas
matplotlib
open3d==0.12.0
catkin_pkg
pybullet==2.7.9
torch==1.7.0
pytorch-ignite
tensorboard
tqdm
trimesh
urchin
cython
scikit-image
pykdtree
escnn
einops